+++ app/src/main/java/com/android/sergiobelda/gramophone/ui/mylibrary/albums/AlbumsFragment.kt
@@ -11,6 +11,7 @@ import androidx.recyclerview.widget.GridLayoutManager
 
 import com.android.sergiobelda.gramophone.R
 import com.android.sergiobelda.gramophone.model.Album
+import com.android.sergiobelda.gramophone.viewmodel.mylibrary.AlbumsViewModel
 import kotlinx.android.synthetic.main.fragment_albums.*
 
 /**
