--- app/src/main/java/com/android/sergiobelda/gramophone/ui/mylibrary/pages/TracksFragment.kt
@@ -1,55 +1,0 @@
-package com.android.sergiobelda.gramophone.ui.mylibrary.pages
-
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.LinearLayoutManager
-
-import com.android.sergiobelda.gramophone.R
-import com.android.sergiobelda.gramophone.model.Track
-import com.android.sergiobelda.gramophone.ui.MainActivity
-import kotlinx.android.synthetic.main.fragment_songs.*
-
-/**
- * A simple [Fragment] subclass.
- *
- */
-class TracksFragment : Fragment() {
-    lateinit var trackSelectedListener: TrackSelectedListener
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_songs, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        recyclerView.layoutManager = LinearLayoutManager(context)
-        //val song = Track("Money", "Pink Floyd")
-        //val song1 = Track("Stairway to Heaven", "Led Zeppelin")
-        val songs = ArrayList<Track>()
-        //songs.add(song)
-        //songs.add(song1)
-        val adapter = TracksAdapter(songs, object : TracksAdapter.OnItemClickListener {
-            override fun onItemClick(track: Track) {
-                trackSelectedListener.onSongSelected(track)
-                (activity as MainActivity).expandBottomSheet()
-            }
-        })
-        recyclerView.adapter = adapter
-    }
-
-    interface TrackSelectedListener {
-        fun onSongSelected(track: Track)
-    }
-}
